requirements.txt
azure-cognitiveservices-speech
simpleaudio
